Move common controller setup method to its own file
